refactor(editor): single shared tiptap base for editors and server rendering
Move StyledHeading and the shared extension set (StarterKit options,
TextAlign, Image, Link) into @op/common/tiptap as buildSharedTiptapBase.
@op/ui's buildBaseExtensions composes it with the client-only extensions
(DetailsFocus, Details node views); serverExtensions composes it with
the schema-only server stubs — deleting its drifted private StyledHeading
copy. Adds a schema-superset test guaranteeing serverExtensions covers
every node/mark the editors can produce (the silent-content-drop guard).

New workspace dependency: @op/ui -> @op/common.
