+++ packages/ui/package.json
@@ -100,6 +100,7 @@
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
     "@internationalized/date": "catalog:",
+    "@op/common": "workspace:*",
     "@op/core": "workspace:*",
     "@op/hooks": "workspace:*",
     "@op/styles": "workspace:*",
